refactor: change return type of some functions

+++ app/src/main/java/com/luisenricke/simpleroomapp/data/dao/ContactDAO.kt
@@ -13,7 +13,7 @@ abstract class ContactDAO : Base<Contact>,
     Base.PrimaryKeyDAO<Contact> {
 
     @Query("SELECT COUNT(*) FROM ${SCHEMA.TABLE}")
-    abstract override fun count(): Int
+    abstract override fun count(): Long
 
     @Query("SELECT * FROM ${SCHEMA.TABLE}")
     abstract override fun get(): List<Contact>
@@ -22,16 +22,16 @@ abstract class ContactDAO : Base<Contact>,
     abstract override fun drop()
 
     @Query("SELECT * FROM ${SCHEMA.TABLE} WHERE id = :id")
-    abstract override fun get(id: Int): Contact
+    abstract override fun get(id: Long): Contact
 
     @Query("SELECT * FROM ${SCHEMA.TABLE} WHERE id IN(:ids)")
-    abstract override fun get(ids: IntArray): List<Contact>
+    abstract override fun get(ids: LongArray): List<Contact>
 
     @Query("DELETE FROM ${SCHEMA.TABLE} WHERE id = :id")
-    abstract override fun delete(id: Int): Int
+    abstract override fun delete(id: Long): Int
 
     @Query("DELETE FROM ${SCHEMA.TABLE} WHERE id IN(:ids)")
-    abstract override fun deletes(ids: IntArray): Int
+    abstract override fun deletes(ids: LongArray): Int
 
     @Query("SELECT * FROM ${SCHEMA.TABLE} ORDER BY ${SCHEMA.ID} LIMIT 1")
     abstract fun last(): Contact?
